\name{ex1}
\alias{ex1}
\alias{ex2}
\alias{ex3}
\docType{data}
\title{Cross-hybridisation example}
\description{Cross-hybridisation example}
\usage{
data(ex1)
data(ex2)
data(ex3)
}
\format{

These objects contain the data of three examples of cross-hybridisation.
}

\references{
Casneuf, Van de Peer and Huber. Submitted.
}

\examples{

   data(ex1)
   data(ex2)
   data(ex3)

   plotExample(ex1)
   plotExample(ex2)
   plotExample(ex3)


## see also the vignette
}
\author{Tineke Casneuf \email{tine@ebi.ac.uk}}
\keyword{datasets}