\name{plotExample}
\alias{plotExample}
\title{plot Cross-hybridisation examples}
\description{Cross-hybridisation examples}
\usage{

\S4method{plotExample}{XhybExamples}()

data(ex1)
data(ex2)
data(ex3)
}
\format{

  \code{plotExample} is an S4 method to plot objects of 'XhybExamples' objects.
}
\references{
Casneuf, Van de Peer and Huber. Submitted.
}

\examples{

data(ex1)
data(ex2)
data(ex3)
plotExample(ex1)
plotExample(ex2)
plotExample(ex3)


## see also the vignette
}

\author{Tineke Casneuf \email{tine@ebi.ac.uk}}
\keyword{methods}

