\name{runSimulation}
\alias{runSimulation}
\title{Cross-hybridisation simulation}
\description{Cross-hybridisation simulation}
\usage{
  runSimulation()

}
\format{

   \code{runSimulation} is a function that will run the simulation we ran for our study.
   By modifying the parameters, users can play with it themselves.
}
\references{
Casneuf, Van de Peer and Huber (Submitted).
}

\examples{
  runSimulation()

## see also the vignette
}

\author{Tineke Casneuf \email{tine@ebi.ac.uk}}
\keyword{manip}